Expand PostgreSQL integration tests and adjust sample data
Expanded the existing SQL Server integration test coverage to include PostgreSQL.

Changed the sample book review value from float to decimal to avoid precision issues that surfaced during PostgreSQL integration testing. This keeps the test data behavior more consistent across database providers.

Adjusted one PostgreSQL integration test to account for PostgreSQL LIKE behavior. Unlike the current SQL Server test behavior, PostgreSQL LIKE is case-sensitive by default.

All new PostgreSQL integration tests are passing except one. Added a TODO note to the remaining failing test because PostgreSQL uses a different symbol for XOR than SQL Server. Fixing that test will likely require a moderate refactor to support dialect-specific operator behavior.

+++ CarriganSqlTools/Carrigan.SqlTools.IntegrationTests/Models/BookStats.cs
@@ -4,15 +4,15 @@ public sealed class BookStats
 {
     public int BookId { get; set; }
     public int StockCount { get; set; }
-    public float AverageReview { get; set; }
+    public decimal AverageReview { get; set; }
 
-    public static string CreateTableSql => 
+    public static string CreateTableSql =>
         """
         CREATE TABLE [BookStats]
         (
             [BookId] INT NOT NULL,
             [StockCount] INT NOT NULL,
-            [AverageReview] REAL NOT NULL,
+            [AverageReview] DECIMAL(10,2) NOT NULL,
             CONSTRAINT [PK_BookStats] PRIMARY KEY ([BookId]),
             CONSTRAINT [FK_BookStats_Book] FOREIGN KEY ([BookId]) REFERENCES [Book] ([Id])
         );
